% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_volcano.R
\name{make_volcano}
\alias{make_volcano}
\title{Graficar varios volcano plot a la vez (loop)}
\usage{
make_volcano(indir, figdir, abslogFC = 1, noGeneNames = 20)
}
\arguments{
\item{indir}{Carpeta de entrada, los resultados de la expresion diferencial deben empezar con DE y terminar con .csv.}

\item{figdir}{Carpeta de salida para las figuras}

\item{abslogFC}{Valor de corte de logFoldChange, default=1}

\item{noGeneNames}{Numero de genes que quieres que tengan nombre en el plot, default =20}
}
\value{
Regresa un volcano plot por cada resultado de expresion diferencial y lo guarda en la carpeta
}
\description{
Graficar varios volcano plot a la vez (loop)
}
\details{
Por cada resultado de Expresion diferencial genera un volcano plot
}
\examples{
make_volcano(indir, figdir, abslogFC = 1, noGeneNames = 20)
}
