% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Volcano_function.R, R/plot_volcano.R
\name{plot_volcano}
\alias{plot_volcano}
\title{Obtener informacion para volcano plot}
\usage{
plot_volcano(file_name, figdir, abslogFC = 1, noGeneNames = 20, plotTitle = F)

plot_volcano(file_name, figdir, abslogFC = 1, noGeneNames = 20, plotTitle = F)
}
\arguments{
\item{file_name}{"DE_plotTitle.csv" resultados de la expresion diferencial deben empezar con DE y terminar con .csv.}

\item{figdir}{Carpeta de salida para las figuras}

\item{abslogFC}{Valor de corte de logFoldChange, default=1}

\item{noGeneNames}{Numero de genes que quieres que tengan nombre en el plot, default =20}

\item{plotTitle}{Si quieres agregar o no el titulo a la figura, default = FALSE.}
}
\value{
Genera un volcano plot por cada archivo de entrada

Genera un volcano plot por cada archivo de entrada
}
\description{
Obtener informacion para volcano plot

Obtener informacion para volcano plot
}
\details{
Funcion que permite obtener los datos para generar el volcano plot, dividiendo la informacion
de acuerdo al logfoldchange que tenemos, ademas podemos seleccionar cuantos genes queremos que tengan etiquetas en el plot.

Funcion que permite obtener los datos para generar el volcano plot, dividiendo la informacion
de acuerdo al logfoldchange que tenemos, ademas podemos seleccionar cuantos genes queremos que tengan etiquetas en el plot.
}
\examples{
plot_volcano(file_name, figdir, abslogFC = 1, noGeneNames = 20, plotTitle = F)
plot_volcano(file_name, figdir, abslogFC = 1, noGeneNames = 20, plotTitle = F)
}
